Deletion of users + ID fix

+++ src/main/java/app/persistency/PictureDAO.java
@@ -1,11 +1,10 @@
 package app.persistency;
 
 import app.domain.pictures.Picture;
-import java.math.BigInteger;
 
 /**
  * Interface for pictures.
  * @author jonathan
  */
-public interface PictureDAO extends DataAccessObject<Picture, BigInteger> {
+public interface PictureDAO extends DataAccessObject<Picture, Long> {
 }
